# api-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the release candidate version of the \"Swiss Corporate API\"\
    \ specification.\n"
  version: "1.0.0.2-SNAPSHOT"
  title: "Swiss Corporate API"
  contact:
    name: "SIX Group"
    url: "https://www.six-group.com/"
    email: "swisscorpapi@six-group.com"
host: "tbd.six-group.com"
basePath: "/api/bankingservices/corporate/v1"
...
schemes:
- "https"
paths:
...
  /accounts:
    get:
      tags:
      - "accounts"
      summary: "Retrieve list of authorized accounts"
      operationId: "requestAccountList"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Bearer followed by a base64 encoded OAuth access token"
        required: true
        type: "string"
      - name: "X-CorAPI-Target-ID"
        in: "header"
        description: "ID that identifies the provider (e.g., a financial institution).\
          \ (SCOPE: SIX)"
        required: true
        type: "string"
      - name: "X-CorAPI-Client-ID"
        in: "header"
        description: "ID of the client forwarded to the provider. (SCOPE: FI)"
        required: true
        type: "string"
      - name: "X-Correlation-ID"
        in: "header"
        description: "Unique ID (defined by the caller) which will be reflected back\
          \ in the response."
        required: true
        type: "string"
      - name: "User-Agent"
        in: "header"
        description: "Name and version of the of the Client software"
        required: true
        type: "string"
      responses:
        200:
          description: "Returns a list of authorized accounts."
          headers:
            X-Correlation-ID:
              type: "string"
              description: "Reflects the ID (set by the caller) from the request."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/accountItem"
        400:
          description: "Bad Request - The format of the request was invalid."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        401:
          description: "Unauthorized - Either no token or an invalid (e.g., token\
            \ expired) token was received."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        403:
          description: "Forbidden - A valid OAuth Token was received, but access was\
            \ denied."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        404:
          description: "Not Found - Either the endpoint does not exist or a requested\
            \ resource is not yet available (e.g., account statements)"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        405:
          description: "Method Not Allowed"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        500:
          description: "Internal Server Error"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
      x-swagger-router-controller: "Accounts"
  /accounts/{accountId}:
    get:
      tags:
      - "accounts"
      summary: "Retrieve information about a single specific account"
      operationId: "requestAccountDetails"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Bearer followed by a base64 encoded OAuth access token"
        required: true
        type: "string"
      - name: "accountId"
        in: "path"
        description: "id of account"
        required: true
        type: "string"
        maxLength: 30
      - name: "X-CorAPI-Target-ID"
        in: "header"
        description: "ID that identifies the provider (e.g., a financial institution).\
          \ (SCOPE: SIX)"
        required: true
        type: "string"
      - name: "X-CorAPI-Client-ID"
        in: "header"
        description: "ID of the client forwarded to the provider. (SCOPE: FI)"
        required: true
        type: "string"
      - name: "X-Correlation-ID"
        in: "header"
        description: "Unique ID (defined by the caller) which will be reflected back\
          \ in the response."
        required: true
        type: "string"
      - name: "User-Agent"
        in: "header"
        description: "Name and version of the of the Client software"
        required: true
        type: "string"
      responses:
        200:
          description: "Returns the details of an account."
          headers:
            X-Correlation-ID:
              type: "string"
              description: "Reflects the ID (set by the caller) from the request."
          schema:
            $ref: "#/definitions/accountItem"
        400:
          description: "Bad Request - The format of the request was invalid."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        401:
          description: "Unauthorized - Either no token or an invalid (e.g., token\
            \ expired) token was received."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        403:
          description: "Forbidden - A valid OAuth Token was received, but access was\
            \ denied."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        404:
          description: "Not Found - Either the endpoint does not exist or a requested\
            \ resource is not yet available (e.g., account statements)"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        405:
          description: "Method Not Allowed"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        500:
          description: "Internal Server Error"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
      x-swagger-router-controller: "Accounts"
  /accounts/{accountId}/balance:
    get:
      tags:
      - "accounts"
      summary: "Retrieve account balance information"
      description: "Retrieve account balance information.\n* Returns the intraday\
        \ booked balance (ITBD), if called without a date.\n* Returns the intraday\
        \ booked balance (ITBD), if called for the current date.\n* Returns the closing\
        \ booked balance (CLBD), if called for a past date."
      operationId: "requestAccountBalance"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Bearer followed by a base64 encoded OAuth access token"
        required: true
        type: "string"
      - name: "accountId"
        in: "path"
        description: "id of account"
        required: true
        type: "string"
        maxLength: 30
      - name: "date"
        in: "query"
        description: "The date to query, formatted as yyyy-mm-dd."
        required: false
        type: "string"
        default: "the current date"
        format: "date"
      - name: "X-CorAPI-Target-ID"
        in: "header"
        description: "ID that identifies the provider (e.g., a financial institution).\
          \ (SCOPE: SIX)"
        required: true
        type: "string"
      - name: "X-CorAPI-Client-ID"
        in: "header"
        description: "ID of the client forwarded to the provider. (SCOPE: FI)"
        required: true
        type: "string"
      - name: "X-Correlation-ID"
        in: "header"
        description: "Unique ID (defined by the caller) which will be reflected back\
          \ in the response."
        required: true
        type: "string"
      - name: "User-Agent"
        in: "header"
        description: "Name and version of the of the Client software"
        required: true
        type: "string"
      responses:
        200:
          description: "Returns account balance information."
          headers:
            X-Correlation-ID:
              type: "string"
              description: "Reflects the ID (set by the caller) from the request."
          schema:
            $ref: "#/definitions/accountBalanceItem"
        400:
          description: "Bad Request - The format of the request was invalid."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        401:
          description: "Unauthorized - Either no token or an invalid (e.g., token\
            \ expired) token was received."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        403:
          description: "Forbidden - A valid OAuth Token was received, but access was\
            \ denied."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        404:
          description: "Not Found - Either the endpoint does not exist or a requested\
            \ resource is not yet available (e.g., account statements)"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        405:
          description: "Method Not Allowed"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        500:
          description: "Internal Server Error"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
      x-swagger-router-controller: "Accounts"
  /accounts/{accountId}/transactions:
    get:
      tags:
      - "accounts"
      summary: "Retrieve transactions of a specific account"
      description: "* Returns the transaction list of the current day, if called without\
        \ a date. * Returns the transaction list for a specific day, if called for\
        \ a past date. In case the specified day has not yet been finalized, the response\
        \ will be a 404 error.\n"
      operationId: "requestAccountTransactions"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Bearer followed by a base64 encoded OAuth access token"
        required: true
        type: "string"
      - name: "accountId"
        in: "path"
        description: "id of account"
        required: true
        type: "string"
        maxLength: 30
      - name: "date"
        in: "query"
        description: "The date to query, formatted as yyyy-mm-dd."
        required: false
        type: "string"
        default: "the current date"
        format: "date"
      - name: "X-CorAPI-Target-ID"
        in: "header"
        description: "ID that identifies the provider (e.g., a financial institution).\
          \ (SCOPE: SIX)"
        required: true
        type: "string"
      - name: "X-CorAPI-Client-ID"
        in: "header"
        description: "ID of the client forwarded to the provider. (SCOPE: FI)"
        required: true
        type: "string"
      - name: "X-Correlation-ID"
        in: "header"
        description: "Unique ID (defined by the caller) which will be reflected back\
          \ in the response."
        required: true
        type: "string"
      - name: "User-Agent"
        in: "header"
        description: "Name and version of the of the Client software"
        required: true
        type: "string"
      responses:
        200:
          description: "Returns transactions of an account."
          headers:
            X-Correlation-ID:
              type: "string"
              description: "Reflects the ID (set by the caller) from the request."
          schema:
            $ref: "#/definitions/accountTransactionReport"
        400:
          description: "Bad Request - The format of the request was invalid."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        401:
          description: "Unauthorized - Either no token or an invalid (e.g., token\
            \ expired) token was received."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        403:
          description: "Forbidden - A valid OAuth Token was received, but access was\
            \ denied."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        404:
          description: "Not Found - Either the endpoint does not exist or a requested\
            \ resource is not yet available (e.g., account statements)"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        405:
          description: "Method Not Allowed"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        500:
          description: "Internal Server Error"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
      x-swagger-router-controller: "Accounts"
...
definitions:
...
  paymentIbanAccount:
    type: "object"
    required:
    - "identification"
    - "type"
    properties:
      type:
        type: "string"
        description: "allowed account identification type for the creditor account\
          \ depends on the payment type. The debtor account must always be an IBAN."
        enum:
        - "IBAN"
      identification:
        type: "string"
        example: "CH9300762011623852957"
        maxLength: 34
        pattern: "[A-Z]{2,2}[0-9]{2,2}[a-zA-Z0-9]{1,30}"
    title: "Payment Account"
    example:
      identification: "CH9300762011623852957"
      type: "IBAN"
...
  balanceCurrencyAmount:
    type: "object"
    required:
    - "amount"
    - "currency"
    properties:
      currency:
        type: "string"
        example: "CHF"
      amount:
        type: "string"
        example: "10.25"
        maxLength: 18
        pattern: "-?[0-9]{1,12}([.][0-9]{1,5})?"
    title: "Balance Currency-Amount"
    example:
      amount: "10.25"
      currency: "CHF"
...
  accountItem:
    type: "object"
    required:
    - "account"
    - "currency"
    - "designation"
    - "id"
    properties:
      id:
        type: "string"
        example: "550e8400-e29b-11d4-a716-446655440000"
      account:
        $ref: "#/definitions/paymentIbanAccount"
      currency:
        type: "string"
        example: "CHF"
      designation:
        type: "string"
        example: "Firmenkonto"
        maxLength: 140
      _links:
        $ref: "#/definitions/accountItem__links"
    title: "Account Information Item"
    example:
      _links:
        balance: "/accounts/550e8400-e29b-11d4-a716-446655440000/balance"
        self: "/accounts/550e8400-e29b-11d4-a716-446655440000"
        transactions: "/accounts/550e8400-e29b-11d4-a716-446655440000/transactions"
      currency: "CHF"
      id: "550e8400-e29b-11d4-a716-446655440000"
      designation: "Firmenkonto"
      account:
        identification: "CH9300762011623852957"
        type: "IBAN"
  accountBalanceItem:
    type: "object"
    required:
    - "balance"
    - "balanceType"
    - "date"
    properties:
      date:
        type: "string"
        format: "date"
        example: "2010-02-15"
      balanceType:
        type: "string"
        example: "CLBD"
        enum:
        - "ITBD"
        - "CLBD"
      balance:
        $ref: "#/definitions/balanceCurrencyAmount"
      _links:
        $ref: "#/definitions/accountBalanceItem__links"
    title: "Account Balance Item"
    example:
      date: "2010-02-15"
      balanceType: "CLBD"
      balance:
        amount: "10.25"
        currency: "CHF"
      _links:
        self: "/accounts/550e8400-e29b-11d4-a716-446655440000/balance"
        transactions: "/accounts/550e8400-e29b-11d4-a716-446655440000/transactions"
        account: "/accounts/550e8400-e29b-11d4-a716-446655440000"
...
  commonErrorResponse:
    type: "object"
    properties:
      type:
        $ref: "#/definitions/commonErrorType"
      title:
        type: "string"
        example: "This is the general problem description"
      detail:
        type: "string"
        example: "Detailed problem description with respect to the current request,\
          \ e.g., invalid account number format"
      instance:
        type: "string"
        example: "path/to/corresponding/resource"
    title: "Common Error Response"
  commonErrorType:
    type: "string"
    title: "Common Error Type"
    description: "Error Types for commonErrorResponse. See Appendix for detailed information."
    example: "/problems/TECHNICAL_ERROR"
    enum:
    - "/problems/INVALID_PAYLOAD"
    - "/problems/MALFORMED_PAYLOAD"
    - "/problems/INVALID_TOKEN"
    - "/problems/EXPIRED_TOKEN"
    - "/problems/INSUFFICIENT_PRIVILEGES"
    - "/problems/NO_ACCESS_TO_RESOURCE"
    - "/problems/RESOURCE_DOES_NOT_EXIST"
    - "/problems/RESOURCE_NOT_READY"
    - "/problems/RESOURCE_TOO_LARGE"
    - "/problems/WRONG_METHOD"
    - "/problems/OPERATION_NOT_ALLOWED"
    - "/problems/TECHNICAL_ERROR"
...
  accountItem__links:
    properties:
      self:
        type: "string"
        example: "/accounts/550e8400-e29b-11d4-a716-446655440000"
      balance:
        type: "string"
        example: "/accounts/550e8400-e29b-11d4-a716-446655440000/balance"
      transactions:
        type: "string"
        example: "/accounts/550e8400-e29b-11d4-a716-446655440000/transactions"
    description: "Contains the paths to additional resources for specific account\
      \ (e.g. path to transactions resource)"
    example:
      balance: "/accounts/550e8400-e29b-11d4-a716-446655440000/balance"
      self: "/accounts/550e8400-e29b-11d4-a716-446655440000"
      transactions: "/accounts/550e8400-e29b-11d4-a716-446655440000/transactions"
  accountBalanceItem__links:
    properties:
      self:
        type: "string"
        example: "/accounts/550e8400-e29b-11d4-a716-446655440000/balance"
      account:
        type: "string"
        example: "/accounts/550e8400-e29b-11d4-a716-446655440000"
      transactions:
        type: "string"
        example: "/accounts/550e8400-e29b-11d4-a716-446655440000/transactions"
    description: "contains the paths to additional resources for specific account\
      \ (e.g. path to transactions resource)"
    example:
      self: "/accounts/550e8400-e29b-11d4-a716-446655440000/balance"
      transactions: "/accounts/550e8400-e29b-11d4-a716-446655440000/transactions"
      account: "/accounts/550e8400-e29b-11d4-a716-446655440000"
...